pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arcanum"
version = "0.1.0"
description = "A simple and secure file encryption and decryption utility."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Bodo Schönfeld", email = "bodo.schoenfeld@niftycode.de" }
]
urls = {"Homepage" = "https://github.com/niftycode/arcanum"}

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython"
]

dependencies = [
    "flake8",
    "black",
    "mypy",
    "pytest",
    "pytest-cov",
    "platformdirs",
    "pycryptodome"
]

[project.scripts]
arcanum = "src.main:main"

[tool.setuptools]
zip-safe = false
packages = { find = { exclude = ["tests"] } }
